a bit of cleaning up + command system (also runClient wont work for some reason before + after the changes, if it fucks anything up im really sorry lol)

+++ src/main/java/us/ferox/client/impl/modules/movement/elytra/mode/NCP.java
@@ -15,27 +15,27 @@ public class NCP extends ElytraMode {
     @Override
     public void onVerticalMovement() {
         if (mc.gameSettings.keyBindJump.isKeyDown())
-            mc.player.rotationPitch = (float) ElytraFlight.ncpRotations.getValue() * -1;
+            mc.player.rotationPitch = ElytraFlight.ncpRotations.getValue().floatValue() * -1;
 
         else if (mc.gameSettings.keyBindSneak.isKeyDown())
-            mc.player.rotationPitch = (float) ElytraFlight.ncpRotations.getValue();
+            mc.player.rotationPitch = ElytraFlight.ncpRotations.getValue().floatValue();
     }
 
     @Override
     public void onHorizontalMovement() {
-        ElytraUtil.accelerateElytra((double) ElytraFlight.horizontalSpeed.getValue());
+        ElytraUtil.accelerateElytra(ElytraFlight.horizontalSpeed.getValue());
     }
 
     @Override
     public void noMovement() {
-        ElytraUtil.freezeElytra(0, (double) ElytraFlight.yOffset.getValue());
+        ElytraUtil.freezeElytra(0, ElytraFlight.yOffset.getValue());
     }
 
     @Override
     public void onRotation() {
-        RotationUtil.lockPitch((double) ElytraFlight.ncpRotations.getValue());
+        RotationUtil.lockPitch(ElytraFlight.ncpRotations.getValue());
 
         if (ElytraFlight.lockRotation.getValue())
-            RotationUtil.lockYaw((double) ElytraFlight.ncpRotations.getValue());
+            RotationUtil.lockYaw(ElytraFlight.ncpRotations.getValue());
     }
 }
